Removing vast swathes of dead code.
